Temporal fix for #364
Temporal fix for #364, now backpacks can be dyed, both belt and pack.
However, GTM chemical dyes can be used for recipes, which is not expected. Besides, backpacks cannot be dyed with chemical bath machines.
The difficulty to fix these bugs is that I have not figured out how to set dyed color, since it's not replacing but mixing with original color.